Update : UserDelete

+++ JAVA_Study/Ex0203/src/view/User/Delete.java
@@ -1,18 +1,22 @@
 package view.User;
  
 import model.MemberDTO;
+import model.UserRoute.AuthDB;
 import model.UserRoute.DeleteDB;
 
 public class Delete {
-	public static boolean deleteService() {
-		boolean result = false;
-		System.out.print("삭제할 유저의 ID : ");
-		String input_id = Input.inputStrig();
+	public static boolean deleteService(MemberDTO member ) {
+		boolean result = false; 
 		System.out.print("삭제할 유저의 PW : ");
 		String input_pw = Input.inputStrig();
-		
-		MemberDTO member = new MemberDTO(input_id, input_pw);
-		result = DeleteDB.deleteDB(member);
+		member.setPw(input_pw);
+		boolean authUser = AuthDB.authDB(member);
+		if(authUser) {			
+			result = DeleteDB.deleteDB(member);
+		}
+		else { 
+			System.out.println("본인 확인에 실패했습니다."); 
+		}	
 		
 		if(result) {
 			//삭제 성공
